addressing the comments
